backend/requirements.txt
flask
flask-cors
flask-limiter
flask-caching
pandas
requests
scikit-learn
python-dotenv
gunicorn